services/agent/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "intelycx-aris-agent"
version = "0.1.0"
description = "Intelycx ARIS WebSocket agent service"
requires-python = ">=3.11"
dependencies = [
  "aiohttp>=3.9.5",
  "python-dotenv>=1.0.1",
  "python-jose>=3.3.0",
  "requests>=2.32.3",
  "pymupdf>=1.24.10",
  "python-docx>=1.1.2",
  "boto3>=1.34.162",
]

[project.scripts]
aris-agent = "app.server.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]



